mail.py: Template env now uses feedparser defined elements
The accessible list of fields is now the same as the list provided as
feedparser. This might be why pubDate was never set. In any case, this
gives much more options.

Additionally, a set of functions accessible from within the the
templates now exists, currently the only accessible function is
time.strftime.